refactor(protocol): rename .inc to .ipp for clangd compatibility
Rename ProtocolMethods.inc to ProtocolMethods.ipp to fix clangd analysis
errors in VSCode. The .ipp extension is a recognized C++ convention for
inline implementation includes.

- Add .clangd config to suppress diagnostics for .ipp files
- Update includes in BitwigProtocol.hpp and Protocol.hpp.template
- Remove obsolete *.inc file association from VSCode settings

+++ src/state/SelectorState.hpp
@@ -40,7 +40,7 @@ using oc::state::SignalVector;
  */
 struct PageSelectorState {
     SignalVector<std::string, MAX_PAGES> names;  // Accumulated cache
-    Signal<int8_t> selectedIndex{0};
+    Signal<int> selectedIndex{0};
     Signal<bool> visible{false};
 
     // Windowed loading state
@@ -83,8 +83,8 @@ struct DeviceSelectorState {
     SignalVector<std::string, MAX_CHILDREN> childrenNames;
     SignalVector<uint8_t, MAX_CHILDREN> childrenTypes;
 
-    Signal<int8_t> currentIndex{0};
-    Signal<int8_t> activeDeviceIndex{0};
+    Signal<int> currentIndex{0};
+    Signal<int> activeDeviceIndex{0};
     Signal<bool> isNested{false};
     Signal<bool> showingChildren{false};
     Signal<bool> showFooter{false};
@@ -138,8 +138,8 @@ struct TrackSelectorState {
     std::array<Signal<bool>, MAX_TRACKS> muteStates{};
     std::array<Signal<bool>, MAX_TRACKS> soloStates{};
 
-    Signal<int8_t> currentIndex{0};
-    Signal<int8_t> activeTrackIndex{0};
+    Signal<int> currentIndex{0};
+    Signal<int> activeTrackIndex{0};
     Signal<bool> isNested{false};
     Signal<bool> visible{false};
